Change packeg from event to movie and names of 2 functions
Changed names of packeges as there names were event, and it was confusing what inside, so I changed it to movie. Olso changes names of function FindOne to FindById for better understanding what function does. Also function UpdateName I changed to find Movie as I update not only name but the whole struct of the movie